changes5

+++ spring-app1/src/main/java/com/alinatkachuk/springtask/dao/DebitCardDAO.java
@@ -14,5 +14,5 @@ public interface DebitCardDAO {
 
     public void editDebitCard(DebitCard debitCard);
 
-    public DebitCard getDebitCardById(int id);
+    public DebitCard getDebitCardById(Long id);
 }
