+++ spring-app1/src/main/java/com/alinatkachuk/springtask/dao/LoanDAO.java
@@ -15,6 +15,6 @@ public interface LoanDAO {
 
     public void editLoan(Loan loan);
 
-    public Loan getLoanById(int id);
+    public Loan getLoanById(Long id);
 }
