+++ spring-app1/src/main/java/com/alinatkachuk/springtask/dao/UserDAO.java
@@ -14,7 +14,7 @@ public interface UserDAO {
 
     public void editUser(User user);
 
-    public User getUserById(int id);
+    public User getUserById(Long id);
 
     public User getUserByEmail(String email);
 }
